requirements.txt
flask
langchain
openai
faiss-cpu
unstructured
python-dotenv
werkzeug